Restructured reusable pieces of code
