+++ src/particle_filter/src/particleFilter.cpp
@@ -410,7 +410,7 @@ bool particleFilter::updateParticles(double cur_M[2][3], vector<vec4x3> &mesh, d
 	}
 	count_bar = 0;
 #endif
-	std::cout << "Distance from the part: " << D << std::endl;
+	// std::cout << "Distance from the part: " << D << std::endl;
 	for (int j = 0; j < cdim; j++) {
 	  particles[i][j] = tempState[j];
 	}
